refactor: make code climate happier
